Move shell integration functionality to pkg/shell
- Move pkg/types/shell.go to pkg/shell/snippets.go
- Move pkg/types/shell_test.go to pkg/shell/snippets_test.go
- Update package declarations and import references
- Update cmd/dodot/root.go to use shellpkg.GetShellIntegrationSnippet
- Update pkg/commands/on/on.go to use shell.GetShellIntegrationSnippet
- All tests pass
